+++ spring/roll_model/src/main/java/com/ccc/roll_model/project/ui/response/GetOpensourceResponse.java
@@ -37,7 +37,7 @@ public class GetOpensourceResponse {
         private Double rSquared;
         private String target;
         private int dataCount;
-        private int runningDuration;
+        private Double runningDuration;
         private int likeCount;
         private int downloadCount;
         private boolean likeYn;
